Modify schema so that Tweets.DateEntry is an int, not a Decimal

+++ constants.py
@@ -40,7 +40,7 @@ class DynamoDBTable(Enum):
             "Id": str,
             "TweetId": str,
             "Date": str,
-            "DateEntry": Decimal,
+            "DateEntry": int,
             "Word": str,
         })
